+++ backend/src/main/java/com/packages/backend/user/UserController.java
@@ -3,6 +3,8 @@ package com.packages.backend.user;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -26,30 +28,84 @@ public class UserController {
   @GetMapping("/user/all")
   public ResponseEntity<List<User>> getAllUsers() {
     List<User> users = userService.findAllUsers();
+    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+    String currentUserEmail = authentication.getName();
+    User connectedUser = userService.findUserByEmail(currentUserEmail);
+    for (User user: users) {
+      if (user.getMessagesReceived() != null) {
+        for (int i=0; i<user.getMessagesReceived().size(); i++) {
+          if (!connectedUser.getId().equals(user.getMessagesReceived().get(i).getFkSender())) {
+            user.getMessagesReceived().remove(i);
+          }
+        }
+      }
+      if (user.getMessagesSended() != null) {
+        for (int i=0; i<user.getMessagesSended().size(); i++) {
+          if (!connectedUser.getId().equals(user.getMessagesSended().get(i).getFkReceiver())) {
+            user.getMessagesSended().remove(i);
+          }
+        }
+      }
+      user.setPassword(null);
+      user.setTokens(null);
+      user.setUserRole(UserRole.HIDDEN);
+    }
     return new ResponseEntity<>(users, HttpStatus.OK);
   }
 
   @GetMapping("/user/find/email/{email}")
   public ResponseEntity<User> getUserByEmail(@PathVariable("email") String email) {
-    User user = userService.findUserByEmail(email);
-    return new ResponseEntity<>(user, HttpStatus.OK);
+    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+    String currentUserEmail = authentication.getName();
+    if (currentUserEmail.equals(email)) {
+      User user = userService.findUserByEmail(email);
+      return new ResponseEntity<>(user, HttpStatus.OK);
+    }
+    else {
+      return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+    }
   }
 
   @GetMapping("/user/find/id/{id}")
   public ResponseEntity<User> getUserById(@PathVariable("id") Long id) {
-    User user = userService.findUserById(id);
-    return new ResponseEntity<>(user, HttpStatus.OK);
+    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+    String currentUserEmail = authentication.getName();
+    User connectedUser = userService.findUserByEmail(currentUserEmail);
+    if (connectedUser.getId().equals(id)) {
+      User user = userService.findUserById(id);
+      return new ResponseEntity<>(user, HttpStatus.OK);
+    }
+    else {
+      return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+    }
   }
 
   @PutMapping("/user/update")
   public ResponseEntity<User> updateUser(@RequestBody User user) {
-    User updateUser = userService.updateUser(user);
-    return new ResponseEntity<>(updateUser, HttpStatus.OK);
+    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+    String currentUserEmail = authentication.getName();
+    User connectedUser = userService.findUserByEmail(currentUserEmail);
+    if (connectedUser.getId().equals(user.getId())) {
+      user.setEmail(currentUserEmail);
+      user.setUserRole(connectedUser.getUserRole());
+      User updateUser = userService.updateUser(user);
+      return new ResponseEntity<>(updateUser, HttpStatus.OK);
+    }
+    else {
+      return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+    }
   }
 
   @DeleteMapping("/user/delete/{email}")
   public ResponseEntity<?> deleteUser(@PathVariable("email") String email) {
-    userService.deleteUserByEmail(email);
-    return new ResponseEntity<>(HttpStatus.OK);
+    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+    String currentUserEmail = authentication.getName();
+    if (currentUserEmail.equals(email)) {
+      userService.deleteUserByEmail(email);
+      return new ResponseEntity<>(HttpStatus.OK);
+    }
+    else {
+      return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+    }
   }
 }
